+++ R/geom_contour_fill.R
@@ -15,10 +15,9 @@
 #' @param bins Number of evenly spaced breaks.
 #' @param binwidth Distance between breaks.
 #' @param na.fill How to fill missing values.
-#' @param fill.linear If `TRUE`, interpolation is bilinear, otherwise is done
-#' with bicubic splines.
+#' @param fill.linear
 #'    - `FALSE` for leting the computation fail with no interpolation
-#'    - `TRUE` for imputing missing values with [akima::interpp]
+#'    - `TRUE` for imputing missing values with [approx]
 #'    - A numeric value for constant imputation
 #'    - A function that takes a vector and returns a numeric (e.g. `mean`)
 #' @param xwrap,ywrap vector of length two used to wrap the circulad dimension.
@@ -67,7 +66,7 @@ geom_contour_fill <- function(mapping = NULL, data = NULL,
                          binwidth = NULL,
                          # na.rm = FALSE,
                          na.fill = FALSE,
-                         fill.linear = TRUE,
+                         # fill.linear = TRUE,
                          xwrap = NULL,
                          ywrap = NULL,
                          show.legend = NA,
@@ -86,7 +85,7 @@ geom_contour_fill <- function(mapping = NULL, data = NULL,
             binwidth = binwidth,
             na.rm = FALSE,
             na.fill = na.fill,
-            fill.linear = fill.linear,
+            # fill.linear = fill.linear,
             xwrap = xwrap,
             ywrap = ywrap,
             ...