+++ R/stat_contour2.r
@@ -23,7 +23,7 @@ stat_contour2 <- function(mapping = NULL, data = NULL,
                           binwidth = NULL,
                           na.rm = FALSE,
                           na.fill = FALSE,
-                          fill.linear = TRUE,
+                          # fill.linear = TRUE,
                           xwrap = NULL,
                           ywrap = NULL,
                           show.legend = NA,
@@ -39,7 +39,7 @@ stat_contour2 <- function(mapping = NULL, data = NULL,
         params = list(
             na.rm = na.rm,
             na.fill = na.fill,
-            fill.linear = fill.linear,
+            # fill.linear = fill.linear,
             breaks = breaks,
             bins = bins,
             binwidth = binwidth,
@@ -101,7 +101,7 @@ StatContour2 <- ggplot2::ggproto("StatContour2", ggplot2::Stat,
   compute_group = function(data, scales, bins = NULL, binwidth = NULL,
                            breaks = scales::fullseq, complete = TRUE,
                            na.rm = FALSE, circular = NULL, xwrap = NULL,
-                           ywrap = NULL, na.fill = FALSE, fill.linear = TRUE) {
+                           ywrap = NULL, na.fill = FALSE) {
       setDT(data)
 
       data <- data[!(is.na(y) | is.na(x)), ]
@@ -124,7 +124,7 @@ StatContour2 <- ggplot2::ggproto("StatContour2", ggplot2::Stat,
           }
       }
 
-      data <- .impute_data.m(data, na.fill, fill.linear)
+      data <- .impute_data.m(data, na.fill)
 
       if (!is.null(xwrap)) {
           data <- WrapCircular(data, "x", xwrap)